requirements.txt
argh==0.26.2
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
dramatiq==1.9.0
gevent==20.6.2
greenlet==0.4.16
pathtools==0.1.2
prometheus-client==0.8.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
sqlparse==0.3.1
watchdog==0.8.3
watchdog-gevent==0.1.0
zope.event==4.4
zope.interface==5.1.0
